Move WriteFlusher out of utils into ioutils
